merged New features branch

+++ src/repositories/comment.repository.js
@@ -3,14 +3,34 @@ class CommentRepository {
     this.db = db;
   }
 
-  // Retrieve all comments for a given video by its ID
-  async getVideoComments(id) {
+  getDb() {
+    return this.db;
+  }
+
+  async getCommentById(commentId, userId) {
     try {
-      const comments = await this.db.executeQuery(
-        "SELECT content FROM comments WHERE video_id = ?",
-        [id]
+      const result = await this.db.executeQuery(
+        "SELECT * FROM comments WHERE id = ? AND user_id = ?",
+        [commentId, userId]
       );
-      return comments;
+      return result;
+    } catch (error) {
+      console.error("Failed to get comment:", error);
+      throw error;
+    }
+  }
+
+  // TODO: Get all the comments for a video by its ID and page and limit
+  async getVideoComments(videoId, page, limit) {
+    try {
+      console.log("video id: ", videoId, "page: ", page, " limit: ", limit);
+      const offset = (page - 1) * limit;
+      console.log(offset);
+      const query = `SELECT * FROM comments WHERE video_id = ? LIMIT ${limit} OFFSET ${offset}`;
+      const params = [videoId];
+
+      const result = await this.db.executeQuery(query, params);
+      return result;
     } catch (error) {
       console.error("Failed to get comments:", error);
       throw error;
@@ -18,11 +38,11 @@ class CommentRepository {
   }
 
   // Add a new comment
-  async addComment(user_id, content, video_id) {
+  async addComment(content, video_id, user_id) {
     try {
       const result = await this.db.executeQuery(
-        "INSERT INTO comments (user_id, content, video_id) VALUES (?, ?, ?)",
-        [user_id, content, video_id]
+        "INSERT INTO comments (user_id,video_id,content) VALUES (?, ?, ?)",
+        [user_id, video_id, content]
       );
       return result;
     } catch (error) {
@@ -32,11 +52,11 @@ class CommentRepository {
   }
 
   // Update an existing comment by its ID
-  async updateComment(content, id) {
+  async updateComment(userId, commentId, content) {
     try {
       const result = await this.db.executeQuery(
-        "UPDATE comments SET content = ? WHERE id = ?",
-        [content, id]
+        "UPDATE comments SET content = ? WHERE id = ? AND  user_id = ?;",
+        [content, commentId, userId]
       );
       return result;
     } catch (error) {
@@ -46,11 +66,11 @@ class CommentRepository {
   }
 
   // Delete a comment by its ID
-  async deleteComment(id) {
+  async deleteComment(commentId, userId) {
     try {
       const result = await this.db.executeQuery(
-        "DELETE FROM comments WHERE id = ?",
-        [id]
+        "DELETE FROM comments WHERE id = ? AND user_id = ?",
+        [commentId, userId]
       );
       return result;
     } catch (error) {
